[None][chore] Move log-probs state and conversions into logprobs.py
Extract the log-probs data types (LogProbsState, LogProbsStateList, the
device-side LogProbsStore) and the three conversion helpers that never
touched the sampler instance: they turn staged host tensors into the
per-request result format and take only their arguments, so they become
module-level functions.

Sizing and the per-step device gather deliberately stay in TorchSampler:
_prepare_log_probs grows TOPK_LOGPROBS_SHAPE and resizes the store in
place, and _process_logprobs reads the sampler-wide shapes, so both are
bound to sampler state rather than to the log-probs data.

SimpleTokenLogprobs / TokenLogprobs / cast move with
get_logprobs_from_request, their only user. Formatting reflows follow from
the reduced indentation. No behavior change.
